Update patch set 2
Patch Set 2:

> Patch Set 2:
> 
> NOTE: The Commit Queue will retry your change automatically.
> 
> The Commit Queue timed out while verifying your change in https://cros-goldeneye.corp.google.com/chromeos/healthmonitoring/buildDetails?buildbucketId=8936121721925633312 . This means that a supporting builder did not finish building your change within the specified timeout. The build failure may have been caused by infrastructure issues, so your change will not be blamed for the failure.
> 
> Commit queue documentation: https://dev.chromium.org/developers/tree-sheriffs/sheriff-details-chromium-os/commit-queue-overview

Actually, a bunch of the failures were on the UnitTest stage for chromeos-ec. Now, it's surprising that unit tests would pass on PreCQ and fail on the CQ, so this might be a flake/false alarm, but can you double-check https://uberchromegw.corp.google.com/i/chromeos/builders/caroline-paladin/builds/4282/steps/UnitTest if the CQ run that started at 4:04p EST fails again in chromeos-ec unit tests?

Patch-set: 2
CC: Gerrit User 1000985 <1000985@3ce6091f-6c88-37e8-8c75-72f92ae8dfba>