charge_manager: allow finding dedicated supply with VOLT > 0 and CUR = 0
This commit is the second tentative after the my last patch got
reverted (original: 7b50f88ba02306bd84f4dea1a9a437e57bfd275b,
revert: 756250daed2ad994fdde6fc162b6fbc61b787a82)

On Fizz if we boot from type-C and later on plug the BJ,
the BJ available current is set to a non-zero value but the current
is set to 0. Because of this it is impossible to detect at runtime
by the kernel that the BJ has been plugged in.

This commit is allowing to find dedicated power supply information
even when the current is equal to 0 as long as the voltage is > 0.
The priority will still be lower than if PWR > 0 or if the supply
is not dedicated.

BUG=chromium:841944
BRANCH=none
TEST=Checked that the kernel can see when the BJ is connected.

Change-Id: Ia932b16c1ec255e145470ea5f6c36e7facf63c6f
Signed-off-by: Fabien Parent <fparent@baylibre.com>