Update patch set 2
Patch Set 2:

How horrible would it be to route reads/writes on the fake I2C address to the vendor command handler?

But it may be better for iteflash to open the other endpoint.  It already has to send a special command to the fake I2C address, so somewhere it's already doing if(using CCD) { }.

Patch-set: 2
Reviewer: Gerrit User 1000187 <1000187@3ce6091f-6c88-37e8-8c75-72f92ae8dfba>