it83xx: read observation register of external timer issue
This bug was opened in the old issue tracker. In the new issue
tracker, the bug link crosbug.com/p/55044 has disappeared and
the bug number can not be passed by pre-uploaded, so the bug
number does not write.

In IT8320 BX version, there is a chance of failure in reading
external timer observation register. We can calculate the
interval of CPU time as a workaround. The bug has been fixed
in the later version of chip.

BUG=none
BRANCH=none
TEST=Ensure the observation register of external timer will
     be read successful in IT8320 DX chip.

Change-Id: I0ec2d0bb83afd1549118de9383dc16cf5adb6b5a
Signed-off-by: tim <tim2.lin@ite.corp-partner.google.com>