ectool: Allow usbpdpower command to take port #
Currently, usbpdpower command enumerates all USB-C ports found
on the board. Barrel jack adapters are not recognized as a USB-C
port, it's not handled by the command.

This patch allows the command to take a port number as a new
parameter.

Signed-off-by: Daisuke Nojiri <dnojiri@chromium.org>

BUG=chromium:841944
BRANCH=none
TEST=Verify ectool usbpdpower 1 return 'Port 1: Disconnected' and
'Port 1: SNK (not charging)' without and with a BJ adapter connected
respectively on Fizz.

Change-Id: If3437c8f3f9952a4cc6004614ccff7bb6bbb4de5