Update patch set 2
Patch Set 2:

> Patch Set 2:
> 
> > By context do you mean you want more text in the commit message, or why the change
> > is required?
> Is this for a Chrome OS project, or is it upstreamed contributions from another project leveraging the codebase? Just curious.

We've been building SDRs that run the chromium-ec together with 'normal' Linux on the AP for a while [1]. I've tried to upstream the board change back then but haven't really heard back on that one, and then kinda dropped the ball... All in all it did work pretty well for us so far.

As I said, this one is for an upcoming product (can't disclose too much, just yet) and I'm just trying to make sure the changes to the core code get merged back in. I realize not all changes make sense for upstream, but the ones that do I intend to send back to you guys.
In general all of our schematics and source code are open so if you're interested feel free
to check it out.

Cheers,
Moritz

[1] https://www.ettus.com/product/details/USRP-N310

Patch-set: 2
Reviewer: Gerrit User 1177024 <1177024@3ce6091f-6c88-37e8-8c75-72f92ae8dfba>