whiskers: increase uart tx buf size to 1K
The formatted error dump is about 400 bytes.  The current 512 bytes TX
buffer might not be enough if we don't read message back immediately.
Increase TX buf size to 1K, this is the largest value possible, and
should be enough if we only read buffer back when there are errors.  (if
we keep polling from USB interface, we can probably use 512 bytes
buffer).

BRANCH=nocturne
BUG=b:112877237
TEST=test on whiskers
Signed-off-by: Wei-Han Chen <stimim@chromium.org>

Change-Id: I4a4840c5c9cb8982c53617a023b9abf67e60f9f0