power_button_x86: Check init-on timeout since PB task start time
When PB state machine is in PWRBTN_STATE_INIT_ON, it first checks if
current time (i.e. time since EC reset) is greater than
CONFIG_POWER_BUTTON_INIT_TIMEOUT (which is typically 1 second). If
yes, then it assumes that it is not okay to bring the AP out of reset
(possibly because battery, charger and PD are not initialized in
time).

However, in case of manual recovery using power and volume buttons,
user could be holding the buttons for > 30 seconds. This is required
to allow user to request recovery with hardware reinitialization. In
such cases, it is not right to check for current time against the
timeout value of 1 second. Instead, the timeout should be considered
since EC started all the tasks.

This change records the timestamp when power button task starts and
then uses that time to calculate the actual timeout in
PWRBTN_STATE_INIT_ON.

BUG=b:113165243
BRANCH=nocturne
TEST=None

Change-Id: Ifd5211cbd1a2fa56c9affcf656df65b11192c519
Signed-off-by: Furquan Shaikh <furquan@google.com>