ec: Replace TPM reset with generic task_reset.
This change is too large to submit, and breaks up neatly into
manageable pieces. Uploading as a single change for the purpose
of discussing the overall approach - and being able to see the
whole picture while doing so.

This adds a generic task_reset function that can reset any task
that declares it supports being reset. Tasks that support reset
must call task_reset_cleanup() on startup, and be prepared to
perform any cleanup / re-initialisation necessary. The task can
control when it can be reset using task_enable_resets() and
task_disable_resets().

A task can request another task be reset at any point; if the
reset cannot happen immediately, it will be queued and will happen
at the earliest point allowed by the task to be reset.

Task resets can be blocking or non-blocking, as specified by
the requestor.

BUG=b:73771117
BRANCH=none
TEST=manual testing on local device

Change-Id: I972184381b005c342374fa16c4dce2ac83e89854
Signed-off-by: Louis Collard <louiscollard@chromium.org>