driver/led/lm3630a: Configure 261ms startup ramp time
It appears that, on startup, LM3630A starts by assuming some high
input PWM value, so when the brightness is set to 0xff, it first
drives the output, and slowly dims it as the PWM input LPF takes
effect (takes about 50ms).

We workaround this issue by adding a 261ms start-up ramp, to
prevent LM3630A from driving backlight right after the brightness
value is set. AFAICT, this has no effect on the user, even when
transitioning from off to on backlight, as the ramp only applies
when the brightness is first set.

BRANCH=nocturne
BUG=b:111010124
TEST=Connect whiskers, no backlight blip

Change-Id: Ia4dafe9a54aa40a4eae865581a06670e94eb6364
Signed-off-by: Nicolas Boichat <drinkcat@chromium.org>