octopus: Factor out common NPCX GPIOs
This commit is a proof-of-concept for splitting out common GPIO
configurations into a shared file in baseboard/. We can see that
configurations for not-connected pins were inadvertently dropped.

I'm open to other ideas; there are likely even better ways to be doing
this. The current requirement for GPIO_INT() to be at the top of the
file makes this implementation a bit awkward.

BUG=b:110192175
TEST=None

Change-Id: I3a53c0354cb252915759604db48f7d8794828db6