cbi: allow tags to be resized
Since we now have a dynamically sized CBI tab (DRAM), it is more
convenient to be able to resize tags. We accomplish this by deleting
the old tag and adding the new tag at the end, after copying all of
the data forward.

BRANCH=none
BUG=b:116075074
TEST=Updated various CBI fields and watch them move to the end. tested
updating the last field and middle fields

Change-Id: Icb9b3cb36445d8e78315f9fd3a74483ff2d52ebf
Signed-off-by: Jett Rink <jettrink@chromium.org>