Update patch set 7
Patch Set 7:

(3 comments)

This is a huge CL and is going to take time to get through.

If we are going through the trouble of making a new statemachine, let's also document at a high-level how things are supposed to work in a md file that we check in as well. Thoughts?

Patch-set: 7
Reviewer: Gerrit User 1259838 <1259838@3ce6091f-6c88-37e8-8c75-72f92ae8dfba>