Update patch set 2
Patch Set 2:

If including a file at a random point is the concern, then here is my random thought.

It is almost based on the idea here. We can make GPIO_INT() to be in their own .inc file and include that file in the gpio.inc file. That way all includes are at the top. One small process we can follow is that include order has to be in alphabetical order - like in Linux kernel. Based on that process, if we can choose our filename appropriately it can get included first.

Patch-set: 2