bc12: Modify bq24392 driver to be a generic BC 1.2 driver
The BC 1.2 USB charger bq24392 chip relies on 2 GPIO signals for control
and determining charge level. There are other BC 1.2 USB charger ICs
which have the same control/feedback signals. This CL changes the
driver name so that it's not specific to this IC, and then can be used
for other BC1.2 chips such as the MAX14637.

This CL doesn't modify any functionality. However, a new flag is added
to indicate the PP5000 control is required as opposed to that being
the default assumption. For the MAX14637, the the control signal to
start BC 1.2 detection also powers the chip.

BUG=b:113267982
BRANCH=none
TEST=make -j buildall

Change-Id: I8c32663b4cb0429aeacde31792d620e7810a0068
Signed-off-by: Scott Collyer <scollyer@google.com>