CHERRY-PICK: nocturne: Set input voltage to 9V when batt full.
To reduce our power consumption in our lowest power state, we should
reduce the charger's input voltage down to 9V when the battery is full
and we are no longer charging it.  This commit will trigger a PD
negotiation to select a 9V source cap.

BUG=b:116125689
BRANCH=firmware-nocturne-10984.B
TEST=Let battery discharge that it will accept current, plug in
blackcat, verify that it chooses a 15V contract.  Wait until battery
gets full and stops accepting current, verify that the new explicit
contract is for 9V.
TEST=While battery is full, turn on AP, verify that contract changes to
15V/3A.

Signed-off-by: Aseda Aboagye <aaboagye@google.com>
Reviewed-on: https://chromium-review.googlesource.com/1239550
Reviewed-by: Furquan Shaikh <furquan@chromium.org>
Commit-Queue: Aseda Aboagye <aaboagye@chromium.org>
Tested-by: Aseda Aboagye <aaboagye@chromium.org>
Change-Id: I8a1ad7b289d4d2a74661658d707c5f9bd44444a9