nocturne: Set input voltage to 9V when batt full.
To reduce our power consumption in our lowest power state, we should
reduce the charger's input voltage down to 9V when the battery is full
and we are no longer charging it.  This commit will trigger a PD
negotiation to select a 9V source cap.

BUG=b:116125689
BRANCH=firmware-nocturne-10984.B
TEST=Let battery discharge that it will accept current, plug in
blackcat, verify that it chooses a 15V contract.  Wait until battery
gets full and stops accepting current, verify that the new explicit
contract is for 9V.

Change-Id: I197f3dc81d693a8afa09f1cefa3aa426eb03971e
Signed-off-by: Aseda Aboagye <aaboagye@google.com>