Update patch set 1
Patch Set 1:

> Patch Set 1: Code-Review-1
> 
> The original BQ24392 driver could only indicate "high" power or "low" power. Does the MAX14637 have the same limitation? Not all BC 1.2 detectors have this limitation and I don't think we should lump them all together. Besides this MAX14637 IC, is there another that behaves exactly the same as the BQ24392?

The MAX14637 IC is a very similar chip and it's connected the same way on DragonEgg. Of the 3 input signals, only 1 is connected. I do have a MAX14637 driver, but it's bq24932.c with all the names changed to max14637. That's why I thought it made sense to have a generic driver that could apply to different chips that are being used.

Patch-set: 1
Reviewer: Gerrit User 1123814 <1123814@3ce6091f-6c88-37e8-8c75-72f92ae8dfba>