Initial checkin of i2c-pseudo Linux kernel module.
This Linux kernel module implements I2C adapters that can be backed by userspace
programs.  This allows for userspace implementations of functionality such as
tunneling I2C through another communication channel to a remote I2C bus.

I intend to submit this driver for inclusion in the upstream kernel.org
repository.  Thus the license and coding style match kernel.org requirements.

<disclaimer>
DO NOT SUBMIT, this is not ready to submit or for a complete review yet!

That said, please look through this and do comment on any significant or high-level issues that you see.  Comments about problems you see with the design, data flow, use of specific kernel interfaces, etc would be welcome!

The things remaining to do are, in no particular order:
- Provide a way for a controller to find its I2C device number.
- Replace use of Linux I2C flag bitmasks with other means (even if still a bitmask) of convenying that information, so that the controller interface isn't dependent on internal Linux I2C bitmask values.
- Change many of the cases where an errno is returned, to instead queuing a more descriptive error message for the controller to read.
- Tests!  At a minimum there should be a simple shell or Python script that one can run as root to exercise most of the functionality, for easy manual verification.  I'll also look into how kernel module tests are typically written & plugged into automation.  (Not implying anything Chrome OS specific, this module has nothing to do with Chrome OS.)
- Some of the pr_warn() are useful messages for a controller author, and will be turned into error messages queued for the controller to read.
</disclaimer>

BRANCH=none
BUG=b:79684405
TEST=Used i2c-tools CLI to i2c-dev to issue I2C queries, and Python interactive interpreter as the I2C pseudo adapter controller.  The i2c-pseudo module correctly proxied write, read, write+read, and multi-byte reads through the controller character device, and it correctly proxied back the responses from the character device, including the specific byte values read.  Later, I implemented an I2C pseudo controller in Servod and use it to flash an ITE EC using both Servo Micro and Suzy-Q.

Change-Id: Idda764e56b9e9d577302b59286e873291ea36d25
Signed-off-by: Matthew Blecker <matthewb@chromium.org>