Update patch set 1
Patch Set 1:

> Patch Set 1:
> 
> Hello, please pardon my noob-ness in tenor and content. I'm a noob at this.
> 
> 
> First, on the "particular prototype" device I experienced b/113207208 , I noticed PS_RDY #2 was never sent. The CC line appeared to go to invalid state -- "vOpen" (>2.67v) -- prior to that. So I think the error is occurring BEFORE the PD_RDY#2. This debouncing fix might not fix it.
> 
> If it is indeed the ChromeOS EC detecting a disconnect, and throwing an Rp (when it should be an Rd), thereby causing "vOpen", I think there might be another root cause. I need to obtain badge access (and equipment permission) to a lab with an oscilloscope and breakout to verify. I currently lack this.
> 
> 
> Second, is the external TCPC debouncing Vbus via hardware? Or are we doing it via software?
> 
> I think the key point is the read time of the ADC. (How long it takes the sampling capacitor to charge, and 5v to actually be read as 5v -- either because of software OR hardware.) If we debounce, it should ONLY be to accommodate the absolute minimum hardware amount -- and not more than that.
> 
> The reason I say this, is as soon as PS_RDY#2 is fired, per the spec, tNewSink = 275ms MAX (t4) later the new-sink can draw up to pSnkSusp = 25mW MAX. (Then it does SRC_CAP => REQUEST => ACCEPT => PS_RDY and can draw current "for real".) True, we can delay more than this for reliability.... but it means the device on the other side is broken.
> 
> 
> 
> For the record, I'm in favor of this change if Aseda is. "Don't let perfect be the enemy of good." Please give me a moment to try to get an oscilloscope and test these changes side-by-side, if only for validation. And thank you for your time.
> 
> Regards,
> Nathan Kolluru
> USB-C Smokejumper and Engineer
> Made for Google Test Team
> 
> 
> 
> From [7.3.10 Source Requested Power Role Swap]:
> 
> The power supply as the new Sink transitions from Swap Standby to drawing pSnkSusp within tNewSnk (t4). The power supply informs the Device Policy Manager that it is operating as the new Sink. At this point subsequent negotiations between the new Source and the new Sink May proceed as normal. The new Sink Shall Not violate the transient load behavior defined in Section 7.2.6 while transitioning to and operating at the new power level. The time duration (t4) depends on the magnitude of the load change.

The debouncing is on the TCPC IC and we don't really have any control over it. Once that IC debounces Vbus, then it will sent and interrupt to the EC that Vbus changed. We need to allow time for that HW debounce to complete before determining that Vbus isn't present and causing us to move to the disconnected state.

Also according to the USB PD spec, Vbus is allowed to be 0 during the power role swap operation, and this debounce is really an extension on that.

Please let me know what you find out with an oscilloscope.  (you can also compare it to the scope capture I have on the attached bug too)

Thank you for looking into this! I appreciate the feedback!

--------

On a tangential note, I think it is worth pointing out that the RX buffer changes most likely exposed this race condition. With the RX buffer changes, we should handle RX events much quicker (which was the main goal), but now we seem to be handling them quicker than the external TCPC debounces VBus, which makes the state machine think we disconnected.

Patch-set: 1
Reviewer: Gerrit User 1259838 <1259838@3ce6091f-6c88-37e8-8c75-72f92ae8dfba>
Label: Trybot-Ready=0