Update patch set 1
Patch Set 1:

> Patch Set 1:
> 
> so, who and when will be setting the serial number? We have some fallbacks for setting board ID if it was not set in the factory process, should setting serial number before that be added?

The idea is that the S/N is only set at factory. And if it's not set there, it should remain undefined/invalid. 
So, if we are setting a board ID at a later time during the fallback workflow, we may decide to ensure S/N is set to some invalid value (can reserve all 0s to be able to reset the S/N, if it was set to something w/o Board ID).

Patch-set: 1
Reviewer: Gerrit User 1152319 <1152319@3ce6091f-6c88-37e8-8c75-72f92ae8dfba>