g: support building with dummy signer
When the image is built on the build server, the cr50-signer is not
available, as the infrastructure required to bring it in does not
exist in this branch.

Note that only board locked RW would require using the real signer,
the release images are signed using ToT, so instead of going through
the trouble of incorporating cr50-utils with all its dependencies into
the cr50-guc-factory-9308.26.B branch let's just have the default
build (ran on the build server) use the local dummy signer, as it did
before, but allow the user to override the signer when producing a
node locked image is required.

BRANCH=none
BUG=none
TEST=verified that both of the following commands succeed:
   emerge-reef chromeos-ec
   make SIGNER=/usr/bin/cr50-codesigner BOARD=cr50 H1_DEVIDS='xx yy' -j

Change-Id: Ida85ac5ffb12ed4895c898f6939909492dc1fcd9
Signed-off-by: Vadim Bendebury <vbendeb@chromium.org>