octopus: motion_sense: provide option to defer the sensor resume actions
It is observed that the lid accelerator not apprears on the i2c bus which
causes the i2c master got NACK in multiple retries during the S5 (and G3)
to S0 power sequence. (i.e. [10106.692955 Lid Accel: 0: init failed: 1])

Learnt from the CL:1091211, refer to the CL:433338 as well, the patch is
similar to the delay option for suspend actions, provide a delay option
for the resume actions as well.

BRANCH=master
BUG=b:116170194
TEST=Examining EC console for KX022 init succeeds in S5->S3->S0 path
on octopus child boards w/ KX022.
(i.e. [10039.971601 Lid Accel: MS Done Init type:0x0 range:4])

Change-Id: Iae4f0fd1e6c46c1b7e12a202a517d8f3b3d82c40
Signed-off-by: Harry Pan <harry.pan@intel.com>