led_pwm: Turn off inactive charge port LEDs.
When a board is configured to only show the charging led on the active
charge port, we should ensure that the other ports' charging LED is
turned off.  This commit fixes a bug where both ports LEDs could remain
lit even though there's only one active charge port at a time.

BUG=b:116020780
BRANCH=firmware-nocturne-10984.B
TEST=Plug in a low power charger on one port, plug in a higher powered
charger on a different port, verify that the charging LED switches to
the higher powered port and the other LED is off.

Change-Id: I48992e36f49346c920170c5bf90761776824af48
Signed-off-by: Aseda Aboagye <aaboagye@google.com>