led_pwm: Fix pulsing bug after init.
When the LED code starts determining what pattern to display, the charge
state machine returns an error briefly at init.  (The error is that the
battery present status is unknown).  Once it determines the battery is
present, it no longer returns an error, but the LED code doesn't cancel
the pulsing pattern.  This commit simply fixes that bug.

BUG=None
BRANCH=firmware-nocturne-10984.B
TEST=Flash nocturne; hibernate; wake via power button, verify that the
LEDs do not pulse red continuously.

Change-Id: Ibe225ba4c136b8758948bd72d9ba688cff15d9e9
Signed-off-by: Aseda Aboagye <aaboagye@google.com>