pdchipinfo: add min firmware version to pdchipinfo
We know that the PS8751 TCPC needs to have firmware version 0x39 or
higher to support properly detecting Vbus presence.

Add a new field to the pdchipinfo host command that exposes the minimum
required firmware version that we know about. This will allow factory
tests or automated test to compare this version to the current version
and fail.

BRANCH=none
BUG=b:116068318
TEST=with corresponding ectool change, min value is reported correctly

Change-Id: Idf338795c3fd6f9f95e51471d0f6a7a422901d52
Signed-off-by: Jett Rink <jettrink@chromium.org>