i2c: Do not attempt bus clear on split xfer failure
If a split transaction fails after its first half, the chip driver has a
few options for dealing with the bus state.  It can either release the
bus by sending a stop bit, or it can leave the bus busy and use a
repeated start for the next transaction.  ChromeEC does not support
multimaster I2C, so either condition is acceptable.

TEST=boot a bip with a battery disconnected.  Observe that the bus reset
error path is not taken when attempting to read the battery
identification registers.
BRANCH=none
BUG=b:116603165, b:112862656

Change-Id: Ieb29233b0701edc2135707247b15a637e9d4d25b
Signed-off-by: Jonathan Brandmeyer <jbrandmeyer@chromium.org>