Update patch set 5
Patch Set 5: Code-Review+2

The only concern I have is how this interacts with DPTF.

Is DPTF not throttling hard enough based on the DRAM temperature?  Or maybe the polling interval for that sensor is too long?

In any event if this helps with ensuring skin temp it is useful...

Patch-set: 5
Reviewer: Gerrit User 1000423 <1000423@3ce6091f-6c88-37e8-8c75-72f92ae8dfba>
Label: Code-Review=+2