sn5s330: Add support for SBU FET control.
The SN5S330 has FETs for the SBU lines which need to be controlled by
the TCPM.  This commit adds a function to control the SBU FETs and
enables them when configuring the USB mux.

This commit also fixes a bug with the external VBUS current limit
setting by setting it appropriately to a min of ~3A.

BUG=b:114340064
BRANCH=firmware-nocturne-10984.B
TEST=Flash nocturne; plug in hoho; verify that external display is
shown. Bounce thru S5, verify that external display still works.
TEST=Repeat above with superspeed hub instead.

Change-Id: Ib28e99d90ba34f0224a5f54a2b79726564a0bde8
Signed-off-by: Aseda Aboagye <aaboagye@google.com>