sn5s330: Add support for SBU FET control.
The SN5S330 has FETs for the SBU lines which need to be controlled by
the TCPM.  This commit adds a function to control the SBU FETs and
enables them when configuring the USB mux.

This commit also fixes a bug with the external VBUS current limit
setting by setting it appropriately to a min of ~3A.

BUG=b:114340064
BRANCH=firmware-nocturne-10984.B
TEST=Flash nocturne; plug in hoho; verify that external display is
shown. Bounce thru S5, verify that external display still works.
TEST=Repeat above with superspeed hub instead.

Change-Id: I931f7a47d4eb28e8d9e3cb188601ce0889a44f8d
Signed-off-by: Aseda Aboagye <aaboagye@google.com>