g: add ITE EC flash programming capability
This patch adds a callback function which triggers generation of the
ITE EC programming sequence by the H1.

The EC programming mode is triggered when the master maintains for at
least 20 ms 200 KHz clock of the SDA line and 100 KHz clock on the SCL
line, locked in phase, with 10% frequency variation tolerance.

To be able to generate a stable 200 KHz clock jitter has to be
disabled, and the actual clock generation function needs to be written
in assembler, as this allows for the most accurate clock frequency
tuning.

The H1 GPIO controller allows to set multiple GPIO pins to requested
values in one 32 bit transaction. The C function maps I2C interface
directly to the GPIOs, prepares the address of the register to use to
control the GPIOs and the values to write to generate all four
combinations of the two signals: 00, 01, 10, 11.

Then it invokes the assembler function to actually generate the
clocks, and then re-enables the clock jitter.

BRANCH=cr50, cr50-mp
BUG=b:75976718
TEST=with the rest of the patches applied verified that it is possible
     to disable and re-enable clock jitter at run time.

Change-Id: Iac33c9bab68fc1ab919d960291176195a08f1791
Signed-off-by: Vadim Bendebury <vbendeb@chromium.org>