Update patch set 2
Patch Set 2:

Sorry if I didn't express this clearly in the previous review, but what's the objection to renaming bq24392.[ch] to max14637.[ch] (and switching all uses of BQ24392 to MAX14637)?

Grunt is currently using bq24392.[ch] for the MAX14637, and doesn't have the 5V rail control. This is achieved via this in gpio.inc:
/* We don't have 5V rail control but the BQ24392 driver expects the GPIO */
UNIMPLEMENTED(EN_PP5000)

I admit this is a little ugly, so perhaps the renamed driver could also be tweaked to support the power options in a cleaner way.

Patch-set: 2
Reviewer: Gerrit User 1212794 <1212794@3ce6091f-6c88-37e8-8c75-72f92ae8dfba>