nocturne: Set up SBU FETs properly.
The SBU FET control should be tied to entering/exiting DP Alt Mode and
not the USB MUX position as the previous commit had.  However, the SBU
lines are also used for CCD and the older boards don't have the
necessary hardware.

BUG=b:114340064
BRANCH=firmware-nocturne-10984.B
TEST=Flash nocturne; verify that external display works after a reboot.
Verify that cr50 is enumerated using SuzyQable.
TEST=Repeat above test on board rev 1.

Change-Id: I0a2422fe70784c3b0e54e5c6cc795c19012f6864
Signed-off-by: Aseda Aboagye <aaboagye@google.com>