bc12: max14637: Add flag to select PP5000 enable control
Some boards need to turn on the PP5000 power rail before starting BC
1.2 detection, but others (such as DragonEgg) don't need PP5000 rail
control tied to BC 1.2 detection. This CL adds a new bit to the flags
so that this control can be board/port specfiic. This flag is set for
all boards currently using this driver as that was the default
behavior.

BUG=b:113267982
BRANCH=none
TEST=make -j buildall

Change-Id: I65d8d1b1a6f6caf6862d3ff4a3f5469bc827b2c8
Signed-off-by: Scott Collyer <scollyer@google.com>