usb_i2c: extend protocol to allow longer writes
It is necessary to be able to send longer than 255 byte packets over
the USB I2C bridge.

This patch introduces a backwards compatible protocol extension to the
protocol to allow that. Namely, the top 4 bits of the byte previously
used as the port number are now used as the top 4 bits of the 12 bit
write counter field.

BRANCH=cr50, cr50-mp
BUG=b:75976718
TEST=with the rest of the patches applied verified that it is possible
     to disable and re-enable clock jitter at run time.

Change-Id: I21475c8d87a6a858ff9ad96dae87388604bbbf2d
Signed-off-by: Vadim Bendebury <vbendeb@chromium.org>