Update patch set 3
Patch Set 3:

I spent some time trying to figure out if it is possible to send both IRQs and the blanket is definitely a tad bit short in all directions. I did end up landing on something which is a bit hacky.

The obvious answer:

	if (mkbp_uses_gpio())
		send_mkbp_event_gpio(active);
        send_mkbp_event_host(active);

suffers from an increase in jitter of O(100us). My hunch is that this is because we are adding extra work for send_mkbp_event() in a timing-sensitive context. I would be happy to look into this more if I am missing something.

The next idea (thanks Furquan) would be to do this from a deferred call. Unfortunately, hook_call_deferred() causes the EC to panic. I ended up diagnosing that this is because we are running with IRQs off, and this causes a crash somewhere in the task switching logic. I didn't delve deeper.

What I landed on is sending the GPIO event, enabling IRQs, and then scheduling the deferred call to send the host event. This works and - while it looks hacky (and may be racy) - we may decide it hardly matters since
a) it's board-specific;
b) we're mostly doing this to support physical button presses in depthcharge (the kernel ignores the eSPI events, so even if they end up racing, it should not matter).

Patch-set: 3