Update patch set 4
Patch Set 4:

This is really hard to review, because you copy-paste the code (but then you remove the extra copy in CL:1235482).

Is there a way to reorder this so that you can avoid the copy? (i.e. just update the code to the new types)

Patch-set: 4
Reviewer: Gerrit User 1129199 <1129199@3ce6091f-6c88-37e8-8c75-72f92ae8dfba>