octopus: remove CONFIG_SYSTEM_UNLOCKED option
For all NPCX boards, remove the CONFIG_SYSTEM_UNLOCKED option. The main
affect of this will be to prevent PD communication in RO. We are
starting to get into factory test that want to verify this behavior. As a
developer, one must remember to jump to RW before PD will negotiate.

BRANCH=none
BUG=b:117109816
TEST=flash on fleex and RO is not negotiate but RW did.

Change-Id: I2df575baad95a334a4a57a9f81600fc039157fd5
Signed-off-by: Jett Rink <jettrink@chromium.org>