iteflash: do not erase by default
It could be beneficial to be able to write flash without erasing it
explicitly, for instance when programming a brand new chip, or when
testing modifications of the iteflash utility.

This patch makes sure that flash is erased only if '-e' command line
parameter is present.

The only place this utility is used by a script in Chrome OS codebase
is flash_ec, it is being updated to maintain current behavior.

BRANCH=none
BUG=none
TEST=verified that running

   $ make BOARD=dragonegg -j
   $ ./util/flash_ec --board=dragonegg

   still succeeds when servod is running and the DUT is connected
   through the servo-v2 board.

Change-Id: Ic238b895ff3324cf8c646908610ea3103d2b78c7
Signed-off-by: Vadim Bendebury <vbendeb@chromium.org>