octopus: enable a1 redriver in S3+
With the upcoming hardware changes, we need to ensure that TypeA port 1
has power during S3+. Previously other hardware signals controlled the
power.

We still want to optimize this signal to turn off power in S3/S0ix if we
know that nothing is plugged in the port.

BRANCH=none
BUG=b:111406013
TEST=flashed on meep without issue, however the actually GPIO toggle is
still untested since we don't have hardware that needs this yet.

Change-Id: I99c548c317a3ec77fef8ece0cc710d072d5b862e
Signed-off-by: Jett Rink <jettrink@chromium.org>