usb-pd: allow a task to prevent TCPC LPM
If a particularly low priority task (like hooks) wants to access the
TCPC, then we do not want the LPM debounce to trigger in the
middle of the communication sequence. This is especially a concern if
the TCPC access is on debug registers that do not push out the LPM
debounce deadline.

BRANCH=none
BUG=b:111406013
TEST=TCPC communication on meep is much more reliable with this change.
TCPC will still go into low power mode after all tasks stop preventing
LPM.

Change-Id: I58cee8e202ced4085f131ff86dbda9d366e1dcca
Signed-off-by: Jett Rink <jettrink@chromium.org>