kukui: Flush SPI FIFO with 0xff before sending data
For some unclear reason, SPI+DMA sometimes fails to transmit the
first few bytes after it is activated the third time, that is,
after the transmission is aborted twice (this looks like a
hardware bug).

To work around this, we flush the FIFO before sending eMMC ack:
this is done by adding a few 0xff at the beginning of the buffer.

Also, print transfer try count to the console, to facilitate
debugging.

BRANCH=none
BUG=b:117253718
TEST=Put kukui in a reboot loop, ~1000 sucessful boots without
     ever requiring a 4th transfer.

Change-Id: I7e5f6bc4da45d9a70ad18447ffb15afd1b62a7a2
Signed-off-by: Nicolas Boichat <drinkcat@chromium.org>