Update patch set 8
Patch Set 8:

> Patch Set 8:
> 
> > Patch Set 8:
> > 
> > > To answer Gwendal's question: I don't think this interrupt will be marked as a wake capable interrupt on the host and so it should not wake the system up from S0ix.
> > 
> > Sigh, that would not be good for us. We're trying to take all measures to make sure the AP runs that one get_time call the moment the gpio gets asserted (without random delays). Having that interrupt not be wake capable defeats a lot of our work.
> 
> Is there a way to mark "wake capable" on the Linux side? Or is it a property of the hardware?

It's a property of the hardware. Why is it necessary to get timestamps on the wake path? I'm not understanding that requirement.

It's possible to set this IRQ as wake capable, but the implementation on the irq path is platform dependent. On the intel systems that typically requires changing the IRQ to a GpioInt (in ACPI parlance) which essentially does de-multiplexing in the gpio driver (think shared IRQ with dispatch to registered IRQ handlers). i.e. you don't get a dedicated vector and interrupt handler.

Patch-set: 8
Reviewer: Gerrit User 1002124 <1002124@3ce6091f-6c88-37e8-8c75-72f92ae8dfba>